Асихронные экшены для user вынесены в отдельный файл
