+++ src/index.tsx
@@ -4,8 +4,8 @@ import { Provider } from 'react-redux';
 
 import App from './components/app/app';
 import { store } from './store';
-import { checkAuthAction } from './store/user/user.slice';
 import ErrorMessage from './components/login/error-message/error-message';
+import { checkAuthAction } from './store/user/user.api-actions';
 
 store.dispatch(checkAuthAction);
