+++ src/services/process-error-handle.ts
@@ -1,5 +1,5 @@
 import { store } from '../store';
-import { clearErrorAction } from '../store/api-actions';
+import { clearErrorAction } from '../store/user/user.api-actions';
 import { setError } from '../store/user/user.slice';
 
 export const processErrorHandle = (message: string): void => {
