+++ src/store/actions.ts
@@ -1,8 +1,4 @@
 import { createAction } from '@reduxjs/toolkit';
-import { AppRoute, AuthorizationStatus } from '../const';
-
-export const requireAuthorization = createAction<AuthorizationStatus>(
-  'USER/requireAuthorization'
-);
+import { AppRoute } from '../const';
 
 export const redirectToRoute = createAction<AppRoute>('USER/redirectToRoute');
