--- src/store/api-actions.ts
@@ -1,10 +1,0 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-import { TIMEOUT_SHOW_ERROR } from '../const';
-import { setError } from './user/user.slice';
-
-export const clearErrorAction = createAsyncThunk(
-  'cities/clearError',
-  (_arg, { dispatch }) => {
-    setTimeout(() => dispatch(setError(null)), TIMEOUT_SHOW_ERROR);
-  }
-);
