+++ src/store/user/user.slice.ts
@@ -1,16 +1,10 @@
-import { PayloadAction, createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { AxiosInstance } from 'axios';
+import { PayloadAction, createSlice } from '@reduxjs/toolkit';
 import {
-  APIRoute,
-  AppRoute,
   AuthorizationStatus,
   NameSpace
 } from '../../const';
-import { dropToken, saveToken } from '../../services/token';
-import { TAuthData } from '../../types/auth-data';
-import { AppDispatch, State, TUserProcess } from '../../types/state';
-import { TUserData } from '../../types/user-data';
-import { redirectToRoute, requireAuthorization } from '../actions';
+import { TUserProcess } from '../../types/state';
+import { checkAuthAction, loginAction, logoutAction } from './user.api-actions';
 
 const initialState: TUserProcess = {
   authorizationStatus: AuthorizationStatus.Unknown,
@@ -18,52 +12,14 @@ const initialState: TUserProcess = {
   errorMessage: null,
 };
 
-export const checkAuthAction = createAsyncThunk<
-  void,
-  undefined,
-  {
-    dispatch: AppDispatch;
-    state: State;
-    extra: AxiosInstance;
-  }
->('user/checkAuth', async (_arg, { extra: api }) => {
-  await api.get(APIRoute.Login);
-});
-
-export const loginAction = createAsyncThunk<
-  void,
-  TAuthData,
-  {
-    dispatch: AppDispatch;
-    state: State;
-    extra: AxiosInstance;
-  }
->(
-  'user/login',
-  async ({ login: email, password }, { dispatch, extra: api }) => {
-    const {
-      data: { token },
-    } = await api.post<TUserData>(APIRoute.Login, { email, password });
-    saveToken(token);
-    dispatch(requireAuthorization(AuthorizationStatus.Auth));
-    dispatch(redirectToRoute(AppRoute.Main));
-  }
-);
-
-export const logoutAction = createAsyncThunk<
-  void,
-  undefined,
-  { dispatch: AppDispatch; state: State; extra: AxiosInstance }
->('user/logout', async (_arg, { dispatch, extra: api }) => {
-  await api.delete(APIRoute.Logout);
-  dropToken();
-  dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
-});
 
 export const userProcess = createSlice({
   name: NameSpace.User,
   initialState,
   reducers: {
+    requireAuthorization(state, action: PayloadAction<AuthorizationStatus>) {
+      state.authorizationStatus = action.payload;
+    },
     setError(state, action: PayloadAction<string | null>) {
       state.errorMessage = action.payload;
     },
@@ -89,4 +45,4 @@ export const userProcess = createSlice({
   },
 });
 
-export const { setError } = userProcess.actions;
+export const { requireAuthorization, setError } = userProcess.actions;
